pyproject.toml
[project]
name = "socmed-backend-sample"
version = "0.1.0"
description = ""
authors = [
    {name = "manabreakmura",email = "174273321+manabreakmura@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-decouple (>=3.8,<4.0)",
    "fastapi[standard] (>=0.115.8,<0.116.0)",
    "sqlmodel (>=0.0.22,<0.0.23)",
    "psycopg[binary] (>=3.2.4,<4.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "pytest (>=8.3.5,<9.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
